--- a/src/components/Header.jsx
+++ b/src/components/Header.jsx
@@ -26,11 +26,19 @@ export default function Header() {
           sizes="120x120"
           href="./assets/apple-touch-icon.png"
         />
+
         <link
           rel="icon"
           type="image/png"
           sizes="32x32"
           href="./assets/favicon-32x32.png"
+        />
+
+        <link
+          rel="icon"
+          type="image/png"
+          sizes="16x16"
+          href="./assets/favicon-16x16.png"
         />
 
         <link