… public/assets/android-chrome-96x96.png
… public/assets/mstile-150x150.png
… public/assets/site.webmanifest