Since #312 added some pretty dramatic perf losses, we're here now.

## Before

### Build
```bash
Page                                       Size     First Load JS
┌ ○ /                                      5.52 kB         133 kB
├   /_app                                  0 B            72.5 kB
├ ○ /404                                   194 B          72.7 kB
├ ○ /about (369 ms)                        306 kB          434 kB
└ λ /api/status                            0 B            72.5 kB
+ First Load JS shared by all              72.5 kB
  ├ chunks/framework-5f4595e5518b5600.js   42 kB
  ├ chunks/main-01df828e572375b9.js        27.6 kB
  ├ chunks/pages/_app-e2615a2c35ea7004.js  2 kB
  ├ chunks/webpack-5752944655d749a0.js     840 B
  └ css/de39ea5c41c10154.css               3.69 kB

✨  Done in 10.10s.
```
### Lighthouse
![Screen Shot 2022-02-28 at 2 13 29 AM](https://user-images.githubusercontent.com/18254823/155965132-f87b8401-d877-40a0-823d-c5200dc7039b.png)

## After

### Build
```bash
Page                                       Size     First Load JS
┌ ○ /                                      7.5 kB         81.2 kB
├   /_app                                  0 B            73.7 kB
├ ○ /404                                   194 B          73.9 kB
├ ○ /about (306 ms)                        308 kB          381 kB
└ λ /api/status                            0 B            73.7 kB
+ First Load JS shared by all              73.7 kB
  ├ chunks/framework-5f4595e5518b5600.js   42 kB
  ├ chunks/main-01df828e572375b9.js        27.6 kB
  ├ chunks/pages/_app-e2615a2c35ea7004.js  2 kB
  ├ chunks/webpack-69b130f3953288db.js     2.04 kB
  └ css/de39ea5c41c10154.css               3.69 kB

✨  Done in 5.87s.
```
### Lighthouse
![Screen Shot 2022-02-28 at 2 10 32 AM](https://user-images.githubusercontent.com/18254823/155964820-f1a427e7-1040-4f96-b5af-f7831c644745.png)