@dev01d 2022-02-28
#### 📝 Changed routes:


| [**/about**](https://personal-website-etx9k5uyh-dev01d.vercel.app/about) | [**/**](https://personal-website-etx9k5uyh-dev01d.vercel.app/) |
|:-:|:-:|
|<a href="https://personal-website-etx9k5uyh-dev01d.vercel.app/about"><img src="https://images.weserv.nl?url=https://sjc.microlink.io/dUME4zXLFzdMeJ4aA5QmtqO13JfxeDkJFJvcIfZ4-r9bjNTTM2eFYY3FkwhYsVyvrqLireyAPEoJqrlty6yADw.png&w=600" alt="Screenshot of /about" width="300"></a><br /><sup><a href="https://images.weserv.nl?url=https://sjc.microlink.io/pS9C03ModtJbf-10BtheFr5ZW_pUnreWFQx6H_X1eIiXNyDiBW_Xa0XaZyO9cjOppoWBQvfgjXpeai11j5-IAw.png">(view full size)</a> |<a href="https://personal-website-etx9k5uyh-dev01d.vercel.app/"><img src="https://images.weserv.nl?url=https://sjc.microlink.io/kapD6DWgYIv_fBFrrulqnqGl0KKXqMGmGZkw2t5CbmkVGelAdHDpb96voFS4NoVnlF2xT1qFpxxhZ2P_FIWwwA.png&w=600" alt="Screenshot of /" width="300"></a><br /><sup><a href="https://images.weserv.nl?url=https://sjc.microlink.io/nFA8XH0rod5ubkFPEBxWOthtcvfO61ai3jScX5N53MYTQLznFeZnmLxFkJid9w5FWBtC6DEvEyeFGt9iR6UjBw.png">(view full size)</a> |







Commit ebb6df2ac7876c4130fcdd05c0c6c71cc719b7aa (https://personal-website-etx9k5uyh-dev01d.vercel.app).